consumer/requirements.txt
SQLAlchemy
confluent-kafka
fastapi
kafka-python
postgres
requests
uvicorn
Authlib
python-multipart
pyjwt